refactored all form containers, added the functionality to edit existing mandatees

+++ app/components/utils/manage-domains/component.js
@@ -21,9 +21,11 @@ export default Component.extend({
 
 		removeMandate() {
 			const domainToDelete = this.get('selectedDomain');
+			if(!domainToDelete) {
+				return;
+			}
 			domainToDelete.destroyRecord();
 			this.set('selectedDomain', null);
-
 		},
 
 		createMandate() {
